# DF/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/orders/schema": {
      "get": {
        "description": "Retrive the datatype of the orders table",
        "operationId": "ordersController_findSchema",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ordersEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "schema validation",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/orders/get": {
      "get": {
        "description": "Filter all the records from the orders table",
        "operationId": "ordersController_findAllmethod",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records to fetch",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ordersEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Filter the records",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "description": "Read only one records from the orders table",
        "operationId": "ordersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ordersEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the only one record",
        "tags": [
          "ERD API"
        ]
      },
      "patch": {
        "description": "Update the record for the orders table",
        "operationId": "ordersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateordersDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ordersEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update the record",
        "tags": [
          "ERD API"
        ]
      },
      "delete": {
        "description": "Delete the record for the orders table",
        "operationId": "ordersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ordersEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/orders": {
      "get": {
        "description": "Read all the records from the orders table",
        "operationId": "ordersController_findAll",
        "parameters": [
          {
            "name": "trs_creator_email",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_created_date",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "trs_created_by",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_modified_date",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "trs_modified_by",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_next_status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_process_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_access_profile",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_org_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_org_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_role_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_role_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_ps_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_ps_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_sub_org_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_sub_org_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ordersEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      },
      "post": {
        "description": "Create the record for the orders table",
        "operationId": "ordersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateordersDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ordersEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/orders/find/first": {
      "get": {
        "description": "Read first record from the orders table",
        "operationId": "ordersController_findFirst",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ordersEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the first record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/orders/find/last": {
      "get": {
        "description": "Read last record from the orders table",
        "operationId": "ordersController_findLast",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ordersEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the last record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/order_items/schema": {
      "get": {
        "description": "Retrive the datatype of the order_items table",
        "operationId": "order_itemsController_findSchema",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/order_itemsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "schema validation",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/order_items/get": {
      "get": {
        "description": "Filter all the records from the order_items table",
        "operationId": "order_itemsController_findAllmethod",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records to fetch",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/order_itemsEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Filter the records",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/order_items/{order_id}": {
      "get": {
        "description": "Read only one records from the order_items table",
        "operationId": "order_itemsController_findOne",
        "parameters": [
          {
            "name": "order_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/order_itemsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the only one record",
        "tags": [
          "ERD API"
        ]
      },
      "patch": {
        "description": "Update the record for the order_items table",
        "operationId": "order_itemsController_update",
        "parameters": [
          {
            "name": "order_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Updateorder_itemsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/order_itemsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update the record",
        "tags": [
          "ERD API"
        ]
      },
      "delete": {
        "description": "Delete the record for the order_items table",
        "operationId": "order_itemsController_remove",
        "parameters": [
          {
            "name": "order_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/order_itemsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/order_items": {
      "get": {
        "description": "Read all the records from the order_items table",
        "operationId": "order_itemsController_findAll",
        "parameters": [
          {
            "name": "trs_creator_email",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_created_date",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "trs_created_by",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_modified_date",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "trs_modified_by",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_next_status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_process_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_access_profile",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_org_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_org_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_role_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_role_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_ps_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_ps_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_sub_org_grp_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trs_sub_org_code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/order_itemsEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Read all the records",
        "tags": [
          "ERD API"
        ]
      },
      "post": {
        "description": "Create the record for the order_items table",
        "operationId": "order_itemsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Createorder_itemsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/order_itemsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create the record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/order_items/find/first": {
      "get": {
        "description": "Read first record from the order_items table",
        "operationId": "order_itemsController_findFirst",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/order_itemsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the first record",
        "tags": [
          "ERD API"
        ]
      }
    },
    "/order_items/find/last": {
      "get": {
        "description": "Read last record from the order_items table",
        "operationId": "order_itemsController_findLast",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/order_itemsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fetch the last record",
        "tags": [
          "ERD API"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest API",
    "description": "The Nestjs API documentation",
    "version": "0.4",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://192.168.2.57:3010",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ordersEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": "number"
          },
          "order_no": {
            "type": "string",
            "example": "string"
          },
          "customer_id": {
            "type": "number",
            "example": "number"
          },
          "order_date": {
            "format": "date-time",
            "type": "string",
            "example": "date"
          },
          "total_amount": {
            "type": "number",
            "example": "number"
          },
          "notes": {
            "type": "string",
            "example": "string"
          },
          "is_active": {
            "type": "boolean",
            "example": "boolean"
          },
          "v_no": {
            "type": "number",
            "example": "number"
          },
          "order_items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/order_itemsEntity"
            }
          },
          "trs_creator_email": {
            "type": "string",
            "example": "string"
          },
          "trs_created_date": {
            "format": "date-time",
            "type": "string",
            "example": "datetime"
          },
          "trs_created_by": {
            "type": "string",
            "example": "string"
          },
          "trs_modified_date": {
            "format": "date-time",
            "type": "string",
            "example": "datetime"
          },
          "trs_modified_by": {
            "type": "string",
            "example": "string"
          },
          "trs_status": {
            "type": "string",
            "example": "string"
          },
          "trs_next_status": {
            "type": "string",
            "example": "string"
          },
          "trs_process_id": {
            "type": "string",
            "example": "string"
          },
          "trs_access_profile": {
            "type": "string",
            "example": "string"
          },
          "trs_org_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_org_code": {
            "type": "string",
            "example": "string"
          },
          "trs_role_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_role_code": {
            "type": "string",
            "example": "string"
          },
          "trs_ps_grp_code": {
            "type": "string",
            "example": "string"
          },
          "trs_ps_code": {
            "type": "string",
            "example": "string"
          },
          "trs_sub_org_code": {
            "type": "string",
            "example": "string"
          },
          "trs_sub_org_grp_code": {
            "type": "string",
            "example": "string"
          }
        },
        "required": [
          "id",
          "order_no",
          "customer_id",
          "order_date",
          "total_amount",
          "notes",
          "is_active",
          "v_no",
          "trs_creator_email",
          "trs_created_date",
          "trs_created_by",
          "trs_modified_date",
          "trs_modified_by",
          "trs_status",
          "trs_next_status",
          "trs_process_id",
          "trs_access_profile",
          "trs_org_grp_code",
          "trs_org_code",
          "trs_role_grp_code",
          "trs_role_code",
          "trs_ps_grp_code",
          "trs_ps_code",
          "trs_sub_org_code",
          "trs_sub_org_grp_code"
        ]
      },
...
    }
  }
}